Dev jon

+++ ZoranEngine/Physics/3D/Collision/CollisionBucket3DBase.cpp
@@ -7,7 +7,7 @@ int CollisionBucket3DBase::FindObject(CollisionObject3DBase * object)
 	std::vector<CollisionObject3DBase*>::iterator iter = std::find(collisionObjects.begin(), collisionObjects.end(), object);
 	if (iter != collisionObjects.end())
 	{
-		return iter - collisionObjects.begin();
+		return static_cast<int>(iter - collisionObjects.begin());
 	}
 	else return -1;
 }
